Fixed bug with view angle of plane in 2D tab.
The plane was often becoming invisible because the the line of sight of the 2D tab viewer was perpendicular to the normal of the plane, this was apparently due to the "up" direction being along the same axis as the normal. The viewAll method is now being called on the 2D tab every time it is selected, this is to ensure the plane is always visible when selecting the 2D view tab. Some PEP8 changes are also included.